backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-rest==0.8.7
psycopg2-binary==2.9.9
six==1.17.0
sqlparse==0.5.3
djangorestframework==3.15.2
pillow==11.0.0
gunicorn==20.1.0
django-cors-headers==4.6.0